Let actor instances share a base actorkeys dict.
In the case where a single actor is run in multiple instances, these are
named with a trailing _ and instance name. For example, the PFS XCU
actor is run in three dewars per spectrograph, so the instances are
started as 'xcu_r1', etc.

Let all of these instances share the common xcu.py keydict.