Be more careful about misssing/invalid values for FITS cards
Values which are not available (the actor is not connected) get an
empty value and have "NOT CURRENT" appended to their comment

Values which are invalid (nans, or whatever is set as "invalid" in the
keyword dictionary) get an empty value and have "INVALID" appended to
their comment.

Note that the implementation juggles how the concept of an invalid
value is handled in opscore.types, fitsio, and pyfits. pyfits is
slated to be retured (INSTRM-331)

Also note that using FITS-compliant empty values might not be
welcome (INSTRM-354). Deal with that issue under that ticket.